sugar_agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
openai>=1.50.0
tavily-python>=0.3.0
pydantic>=2.0.0
